--- app/src/main/java/com/starters/yeogida/network/ApiClient.kt
@@ -1,33 +1,0 @@
-package com.starters.yeogida.network
-
-import com.starters.yeogida.data.api.UserService
-import okhttp3.OkHttpClient
-import okhttp3.logging.HttpLoggingInterceptor
-import retrofit2.Retrofit
-import retrofit2.converter.gson.GsonConverterFactory
-
-interface ApiClient {
-    companion object {
-        private const val BASE_URL = "http://54.180.100.93:8080/api/v1/"
-
-        private val okHttpClient: OkHttpClient by lazy {
-            OkHttpClient.Builder()
-                .addInterceptor(HttpLoggingInterceptor().apply {
-                    level = HttpLoggingInterceptor.Level.BODY
-                })
-                .build()
-        }
-
-        private val retrofit: Retrofit by lazy {
-            Retrofit.Builder()
-                .baseUrl(BASE_URL)
-                .addConverterFactory(GsonConverterFactory.create())
-                .client(okHttpClient) // 로그캣에서 패킷 내용을 모니터링 할 수 있음 (인터셉터)
-                .build()
-        }
-
-        val userService: UserService by lazy {
-            retrofit.create(UserService::class.java)
-        }
-    }
-}
